Translate ErrorEmptyState
I removed the _DescriptiveError class, because I would need the context inside this class, for calling S.of(context).
All the solutions that I could think for passing the context to the _DescriptiveError and keep the final properties/constructors in the ErrorEmptyState class were complicated. So I removed all that stuff and got a much more simple solution.
In this solution I don't have the flexibility of having two constructors and the generation of content is coupled to the widget. These are not problems now though, so I opted for this more narrowed solution.
There is a bit of code duplication between _getTitle() and _getContent(), but it's not a huge deal.